refactor: ImageMetadataProvider -> Helper 분리
