+++ Projects/Feature/Sources/ShareView.swift
@@ -26,39 +26,3 @@ struct ShareImage: View {
             .frame(width: 200, height: 200)
     }
 }
-
-class ImageMetadataProvider: UIActivityItemProvider {
-  var image: UIImage
-
-  override var item: Any {
-    get {
-      return self.image
-    }
-  }
-
-  override init(placeholderItem: Any) {
-    guard let image = placeholderItem as? UIImage else {
-      fatalError("DEBUG ::: Couldn't create image from provided item")
-    }
-
-    self.image = image
-    super.init(placeholderItem: placeholderItem)
-  }
-
-  @available(iOS 13.0, *)
-  override func activityViewControllerLinkMetadata(_ activityViewController: UIActivityViewController) -> LPLinkMetadata? {
-
-    let metadata = LPLinkMetadata()
-    metadata.title = "Result Image"
-
-    var thumbnail: NSSecureCoding = NSNull()
-    if let imageData = self.image.pngData() {
-      thumbnail = NSData(data: imageData)
-    }
-
-    metadata.imageProvider = NSItemProvider(item: thumbnail, typeIdentifier: "public.png")
-
-    return metadata
-  }
-
-}
